requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
dictop==0.1.4
Django==3.2.7
django-smartfields==1.1.3
idna==3.3
magic-import==0.1.3
mysqlclient==2.1.0
Pillow==8.4.0
PyMySQL==1.0.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7